requirements.txt
selenium
pandas
pandas-stubs
mypy
requests
types-requests
bs4
types-beautifulsoup4
python-dotenv

